manager-api/requirements.txt
Flask==3.0.2
libvirt-python==10.0.0
PyYAML==6.0.1